extract commercial metrics to an empty component
Following discussions with @oliverlloyd,
The component now only re-renders when its internal state changes

This might pave the way for a new way for external teams to contribute
to the dotcom-rendering codebase 🤞
